Added Locations and Passenger Cells
Added some locations to the scene that the player will have to drop passengers off at. Also got a reset button working, but it is far from perfect. Got the cells to work, so you know keep track of all your passengers, their desired location and how long they have been with you.
Next step is getting the passengers in the main scene...